Add files via upload
Input Field: An <input> element with type="number" is added so you can enter the value of relay. It has an id="relayValueInput" to be referenced from JavaScript.
Validation: Validation is added to ensure the input field is not empty before sending data to Ubidots. If it is empty, an alert is shown, and the execution of the enviarDato() function is halted.
Function enviarDato(): Modified to get the value of relay from the input field relayValueInput using document.getElementById(). Then, it proceeds to make the POST request to Ubidots with that value.
Error Handling: An additional check is added within xhr.onreadystatechange to handle errors in the POST request. If the request status is not 201, an error message is displayed in the console.